requirements.txt
requests>=2.31.0
openai>=1.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
rich>=13.0.0
